+++ tools_proj/XCPP/GameCore/runtime/AIRuntimeRotateSelf.cpp
@@ -16,8 +16,8 @@ AIRuntimeRotateSelf::~AIRuntimeRotateSelf()
 void AIRuntimeRotateSelf::Init(AITaskData* data)
 {
 	AIBase::Init(data);
-	SinglemAIArgMax = data->vars["SinglemAIArgMax"]->val.get<double>(); 
-	SinglemAIArgMin = data->vars["SinglemAIArgMin"]->val.get<double>(); 
+	SinglemAIArgMax = (float)data->vars["SinglemAIArgMax"]->val.get<double>(); 
+	SinglemAIArgMin = (float)data->vars["SinglemAIArgMin"]->val.get<double>(); 
 	
 }
