+++ tools_proj/XCPP/GameCore/runtime/AIRuntimeValueDistance.cpp
@@ -16,7 +16,7 @@ AIRuntimeValueDistance::~AIRuntimeValueDistance()
 void AIRuntimeValueDistance::Init(AITaskData* data)
 {
 	AIBase::Init(data);
-	floatmAIArgMaxDistance = data->vars["floatmAIArgMaxDistance"]->val.get<double>(); 
+	floatmAIArgMaxDistance = (float)data->vars["floatmAIArgMaxDistance"]->val.get<double>(); 
 	
 }
