+++ tools_proj/XCPP/GameCore/runtime/AIRuntimeValueHP.cpp
@@ -16,8 +16,8 @@ AIRuntimeValueHP::~AIRuntimeValueHP()
 void AIRuntimeValueHP::Init(AITaskData* data)
 {
 	AIBase::Init(data);
-	Int32mAIArgMaxHP = data->vars["Int32mAIArgMaxHP"]->val.get<double>(); 
-	Int32mAIArgMinHP = data->vars["Int32mAIArgMinHP"]->val.get<double>(); 
+	Int32mAIArgMaxHP = (int)data->vars["Int32mAIArgMaxHP"]->val.get<double>(); 
+	Int32mAIArgMinHP = (int)data->vars["Int32mAIArgMinHP"]->val.get<double>(); 
 	
 }
