+++ tools_proj/XCPP/GameCore/runtime/AIRuntimeWait.cpp
@@ -16,10 +16,10 @@ AIRuntimeWait::~AIRuntimeWait()
 void AIRuntimeWait::Init(AITaskData* data)
 {
 	AIBase::Init(data);
-	floatwaitTime = data->vars["floatwaitTime"]->val.get<double>(); 
+	floatwaitTime = (float)data->vars["floatwaitTime"]->val.get<double>(); 
 	boolrandomWait = data->vars["boolrandomWait"]->val.get<bool>(); 
-	floatrandomWaitMin = data->vars["floatrandomWaitMin"]->val.get<double>(); 
-	floatrandomWaitMax = data->vars["floatrandomWaitMax"]->val.get<double>(); 
+	floatrandomWaitMin = (float)data->vars["floatrandomWaitMin"]->val.get<double>(); 
+	floatrandomWaitMax = (float)data->vars["floatrandomWaitMax"]->val.get<double>(); 
 	
 }
